pyproject.toml
[tool.poetry]
name = "resawod-nubapp"
version = "1.2.0"
description = "Resawod slots auto booking"
authors = ["NeoSlugman <neoslugman@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
python-crontab = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
